+++ proto/client-api/src/lib.rs
@@ -98,6 +98,36 @@ pub enum Command {
         peer_id: PeerId,
         candidate: IceCandidate,
     },
+    /// Web Client sends Peer Connection metrics.
+    AddPeerConnectionMetrics {
+        peer_id: PeerId,
+        metrics: PeerMetrics,
+    },
+}
+
+/// Web Client's Peer Connection metrics.
+#[cfg_attr(feature = "medea", derive(Deserialize))]
+#[cfg_attr(feature = "jason", derive(Serialize))]
+#[cfg_attr(test, derive(Debug, PartialEq))]
+#[derive(Clone)]
+pub enum PeerMetrics {
+    /// Peer Connection's ICE connection state.
+    IceConnectionStateChanged(IceConnectionState),
+}
+
+/// Peer Connection's ICE connection state.
+#[cfg_attr(feature = "medea", derive(Deserialize))]
+#[cfg_attr(feature = "jason", derive(Serialize))]
+#[cfg_attr(test, derive(Debug, PartialEq))]
+#[derive(Clone)]
+pub enum IceConnectionState {
+    New,
+    Checking,
+    Connected,
+    Completed,
+    Failed,
+    Disconnected,
+    Closed,
 }
 
 /// Reason of disconnecting Web Client from Media Server.